requirements.txt
torch>=2.1
torchvision>=0.16
diffusers>=0.24
torchmetrics>=1.2
Pillow>=10.0
tqdm>=4.66
